fix(network): Emit `Event.JOIN_FAILED` (#798)
Fixed event emit logic in `Node#emitJoinFailed`. The event was never emitted.

After that change two tests suites failed: `wait-for-join-subscribe-and-publish.test.ts` and `tracker-counters-are-stream-part-specific.test.ts`.  The reason was that the first `sendStatus` always failed because there was no existing tracker connection, and in that situation we emitted `JOIN_FAILED`. 

Changed the logic so that we don't emit `JOIN_FAILED` in `sendStatus` (it is emitted when the tracker instruction is handled and `subscribeAndWaitForJoin` receives the event from there).

### Other changes

Fixed event signatures of three node events which can may have `nodeId` as `null`  in the event payload.