feat(utils):  Enhance `waitForEvent` and `withTimeout` (#829)
Added event filter support to `waitForEvent`. A predicate function is called for each received event. The waiting resolves when the predicate returns true for some event. 

The new implementation required us to limit `eventName` to be `string` instead of not `string | symbol`.

Added `AbortController` support to `waitForEvent` and `withTimeout`.

### Future improvements

The type information for `AbortController` does not yet have definition for `AbortController#signal`. When the type information is available, we can remove "as any" type casting from `withTimeout.ts`.