Groupkey exchange via proxies (#826)
This PR makes it possible for nodes utilizing the proxy pattern to propagate GroupKey exchange related control traffic.

As a new requirement the proxy node must be made aware of the "identity" of the proxied publishers and subscribers. The identity is used by the proxy to route GroupKey related traffic to its appropriate destination.