feat(client): Pre-agreed encryption keys (#839)
If a publisher and a subscriber don't want to exchange the keys via the network, they can use pre-agreed keys like this:
```
const key = new GroupKey('key-id', crypto.randomBytes(32))
publisher.updateEncryptionKey({
    key,
    streamId,
    distibutionMethod: 'rekey'
})
subscriber.addEncryptionKey(key, streamId)
```

This way both participants have the group key available before message delivery starts. For subscriber, we just add a key to the local database. For publisher it we also mark the key to be the `currentGroupKey` which ensures that the next message is published with that key. 

Before this PR there was already the `updateEncryptionKey` method. In this PR we added the `addEncryptionKey` method.

Removed also an obsolete `encryptionKeys` config option, which was previously used to implement similar functionality.