feat(client): Limit contract concurrency (#843)
There is a limit for concurrent contract calls. The limit is per contract and is configurable by config option `maxConcurrentContractCalls`.

A new `ContractFactory` helper class was introduced so that e.g. registry components don't need to know the details of our contract configuration options (e.g. by reading `maxConcurrentContractCalls` from the client config and passing that value to `createDecoratedContract()`).