refactor: Remove unused variables (#868)
Fixed most of the warnings about unused variables:
- removed some variables which were not needed
- prefixed some variables with underscore (e.g. positional methods parameters)
- commented-out some variables (if a piece code is under construction at the moment)

Added an `eslint` rule which catches most of these situations. It doesn't catch unused constructor parameters and unused fields, see e.g.:
- `packages/client/src/subscribe/Resends.ts`
- `packages/network-tracker/src/logic/LocationManager.ts`

If we'd get valid error messages by adding these lines to root-level `tsconfig.jest.json`
```js
{
    "noUnusedLocals": true,
    "noUnusedParameters": true
}
```

But it is more convenient that we gather the analysis via `eslint`. Not sure if it is possible to configure `eslint` to catch constructor parameter and field errors, too.