feat: Require signed messages (#876)
`StreamMessage` requires a signature. Non-signed messages are no longer supported. Removed `SignatureType.NONE`, which was previously the default signature type. Now the default is `SignatureType.ETH`.

StreamMessage validates that we pass `signatureType` when we create `StreamMessage`. Therefore if a client would receive a message which is not signed with `ETH` signature type, it would ignore it (it doesn't trigger `sub.on('error')` as the validation fails already on network node level, see `NodeToNode.ts:113`).

As a signature must be passed to `StreamMessage` as a constructor parameter, the signature payload utility function was extracted from `StreamMessage.getPayloadToSign` to a separate utility function outside of `StreamMessage` class.

### Open questions

Maybe we should remove `SignatureType` field altogether? All messages should be signed with that type and we don't currently plan to support other signature types. 

We don't need to modify the field from protocol (so that we can keep backwards-compatibility). But we could remove the field from `StreamMessage`. We could also check that the `signatureType` has a valid value of 2, when we deserialize a `StreamMessage` JSON array.