refactor: Remove `signatureType` (#879)
Removed `signatureType` field from `StreamMessage`. All messages should be signed with that type and we don't currently plan to support other signature types.

We don't need to modify the field from protocol so that we can keep backwards-compatibility.

Any `StreamMessage` which has `signatureType !== 2` (the `ETH` signature type) is considered invalid. Now it is checked explicitly in `StreamMessageSerializerV32`. Previously the `signatureType` the field was checked in `StreamMessage` constructor, and it accepted exactly the same value (as `ETH` was the only valid `signatureType`). Therefore there is no functionality change in this PR.