refactor(client): Simplify `GroupKey` data model (#880)
Removed `GroupKey#hex` field. It contained the data of `GroupKey#data` but in different format.

There is no need to store the data in two different formats. The only component which used the `hex` format was `GroupKeyStore`. But as all other components used `data` format, we always converted the data between the two formats in practice.

Removed also obsolete methods of `GroupKey`.