refactor(client): Stream factory (#895)
Streams are now created  by calling `StreamFactory#createStream`. There is no state handling in the factory: it just stores the  DI dependencies required by a `Stream` constructor and passes those to the constructor when `createStream` is called.

Before this PR the dependencies were injected by passing a `BrubeckContainer` (i.e. a DI container) instance to the Stream constructor. `BrubeckContainer` is no longer needed, and it is now removed.

Changed also Stream DI fields visibility to private.

### Future improvement

Currently we need to prefix the DI field with underscore so that the field are excluded by `Stream#toObject`. Rename the fields, if/when that is no longer needed.