Include protobuf generated files in git (#906)
Include protobuf generated files in git repository and avoid rebuilding each time package is built. Why?
- By not re-generating them before each build, the `npm run build` script can better utilize tsbuild caches. This leads to a faster compilation time when source code hasn't been touched, and is especially impactful when building the entire monorepo at once.
- Generated files don't change that often.
    - Looks like the output is the same until either source definition files or protoc utility flags are changed?
    - On the plus side, the generated code looks rather legible, so it's not like we are adding some binary blobs into git.
    
The downside is we need to remember to run `npm run generate-protoc-code` any time we updated the protobuf definitions. I think this is a reasonable trade-off given that they won't be probably changing all the time.