refactor(client): Hide ResendSubscription from public API (#912)
Changed the public API of `StreamrClient#subscribe` to return a Subscription instead of `Subscription` | `ResendSubscription`. 

The `ResendSubscription` is a subclass of `Subscription`, and are no extra methods in that class. There is just an extra event type `resendComplete`. Of course non-resend subscription won't emit that message, but in the big picture our public API is more understandable if we don't have a separate class for resends.

Moved all event types to `Subscription` class. Now `error` event type is just one of the event types.