feat(client): Log warning in resend subscribe if no storage assigned (#930)
If storage node is not assigned to a stream and we call `client.resendSubscribe`, it logs a warning. Discussed here: https://github.com/streamr-dev/network-monorepo/pull/921#issuecomment-1288784070

## Testing

Added `FakeLogger` test environment class. It stores log entries to an array, and in test we can check that an expected entry exists in that array.

Created a type annotation for `LogLevel` and used that in `FakeLogger` and `StrictStreamrClientConfig`.