refactor(client): `MessageStream` as `AsyncIterable` (#944)
## Summary

The `MessageStream` class does no longer extends `PushPipeline`. This doesn't change the end-user interface of the class: it is still seen as a class implements `AsyncIterable`  and has no other methods. 
- Before this PR the `MessageStream` class implemented `AsyncGenerator` (as it extended `PushPipeline`). Now as it implements just `AsyncIterable`, there is better separation of concerns between the class and the data generator.
- A small difference in API docs is that now there is no non-documented super class included in the class hierarchy

Currently `MessageStream` contains a private field for the pipeline, and there are delegator methods to access the pipeline. This doesn't change the functionality but allows us to encapsulate the `MessageStream` pipeline functionality. This PR is a milestone to allow us to refactor `MessageStrean` so that most (or all) of those delegator method are no longer needed. 
- Instead e.g. `SubscribePipeline` could create a `PushPipeline` instance and control the data producer part of the pipeline locally. It could then give that instance to `MessageStream` as a `Pipeline` or `AsyncIterable`. `MessageStream` would be just a consumer of the pipeline data.

Changed also the return type of these `StreamrClient` methods from `AsyncGenerator` to `AsyncIterable`
- `getStreamPublishers`
- `getStreamSubscribers`
- `searchStreams`

## Other changes

- using collect method from iterator utils instead of `MessageStream#collectContent` and in some use cases where the collect method of `GeneratorUtils` was used
- fixed `Resends2` test case `"decodes resent messages correctly"`, which compared signature fields for content objects (and therefore always compared two `undefined` values)
- as we no longer extend `PushPipeline`, the `@internal` annotations are not needed in `PushPipeline` and in `Pipeline`

## Limitations and future improvements

There is some overlapping with `GeneratorUtils` and iterator utils: e.g. both have `collect` method. Most or all of those methods are transforming some input data (i.e. consuming it), and therefor we should maybe prefer iterator utils methods instead of `GeneratorUtils` methods. The fact that we implement the methods as `AsyncGenerators` is just an implementation details.