refactor(client): Rename message listener (#945)
Renamed message listener interface `SubscriptionOnMessage`/`MessageStreamOnMessage` to `MessageListener`. This is a change to the public API as the type was exported in `SubscriptionOnMessage` in `index-exports.ts` and documented in API docs. But in typically end-users don't specify the the type of the listener, and therefore this doesn't affect many use case.