test(client): Explicit imports (#946)
Do not import from `src/index.ts`. That allows us to easily search dependencies of source files when refactoring.

Also put `reflect-metadata` import to a separate block as it should always be the first import of a file. This way IDE's `organize import` tool doesn't change the ordering of that import statement.