fix(client): Clean up `Subscription` event listeners (#927)
## Summary

Event listeners assigned to `Subscription` were not cleaned up.

The regression was introduced in https://github.com/streamr-dev/network-monorepo/pull/912. In that PR `eventEmitter` was moved from `ResendSubscription` to `Subscription`, but cleanup logic was not moved. Now cleanup is implemented in `Subscription` super class and therefore we cleanup listeners for both `Subscription` and `ResendSubscription`. 

The cleanup is now part of `Subscription#unsubscribe` logic. Therefore it is done:
- when a subscription is unsubscribed manually
- or automatically when `StreamrClient#destroy` calls that method via `Subscriber`

It ok to do it there as unsubscribed subscriptions don't emit any events. 

## Other changes

- `StreamrClient#destroy` calls `Subscriber#unsubscribe` directly
- Removed unused `OrderMessages#loggerFactory` field (the constructor parameter was used, but the field was not)

## Future improvements

We shouldn't divide the unsubscribe logic to both `Subscription#unsubscribe` and `StreamrClient#unsubscribe`: maybe `StreamrClient#unsubscribe` could just call `Subscription#unsubscribe`.
- A unsubscribed subscription is removed from `Subscriber`, because it attached `sub.onBeforeFinally` listener to the subscription. Maybe `StreamrClient#unsubscribe` could call that directly. (Note that the logic is there also because we automatically unsubscribe streams which break out from the iterator loop. If we'd remove the listener, that feature would not work anymore.)