feat(client)!: Use `Message` instead of `StreamMessage` in public API (#958)
## Summary

This is a breaking change.

`StreamrClient` doesn't use raw `StreamMessage` in the public API of `publish`, `subscribe` and `resend` . Instead it exposes a new interface called Message, which is very similar to `StreamMessage`. That data structure is a flat object which has a subset of `StreamMessage` methods exposed as fields. 

This way it is easier to access the metadata of subscribed/resent messages and we don't need to expose a low level `StreamMessage` class.

In practice:
- changed the 2nd parameter of `onMessage` callback to be `MessageMetadata` instead of `StreamMessage` (`StreamrClient#subscribe` and`StreamrClient#resend`)
- changed iteration type of `Subscribe` and `MessageStream`: the async iterator yields `Message` objects instead of `StreamMessage` objects
- changed return type of publish to `Message` instead of `StreamMessage` `(Stream#publish` and `StreamrClient#publish`)

The `Message` interface doesn't use type generics, i.e. the type of content is always `unknown`. This is different from `StreamMessage` which is usually annotated with the content type `T`. Decided not to use type generics because data consumers can't be sure about the structure of `content` (any type of content can be sent to any stream). Therefore it is good to require type casting when content field is accessed.

We already had an exported interface called `MessageMetadata` , which was used in `StreamrClient#publish` and `Stream#publish` public API.  Renamed the previous interface to `PublishMetadata`. This is not a breaking change as v6 didn't have that exported. In the future these two interface could maybe be merged or somehow connected (e.g. by implementingg`PublishMetadata` by picking some fields from `MessageMetadata`).

Note that the 2nd parameter of `onMessage` callback is `metadata`, not `message`. It is just a the `Message` object, but without `content`. The term metadata was already used in README, but previously in code we referred that as `streamMessage`.

## Other API changes

Modified also `StreamrClient#waitForStorage`. The method parameter is now `Message` instead of `StreamMessage`, and matchFn also operates on `Message` class.
- If there was a timeout error, the error included the `streamMessage` field. It just contained the message object which was given as a method parameter. That field is no longer included there. (The error object is not part of public API).

## Internal usage

Currently `Message` interface has an internal `streamMessage` property. 

We could maybe add a flag to `subscribe` method which would allow it to produce raw `StreamrMessage`s instead of `Message`s if the flag it set (or a separate method). In that scenario there is maybe no need for separate `streamMessage` property. Otherwise we could consider to expose this as a public field (is the field name good?).

Added also internal `getStreamMessages` method to `MessageStream` as some internal components need to iterate the data set using `StreamrMessages`. This could also use the flag mentioned above.

## Remaing usages of StreamMessage in semi-public API

These methods are the remaing public API methods which use `StreamMessage`:
- in `NetworkNodeStub`
- field of `DecryptError` (and other non-exported `Errors` which extend from `DecryptError`
  - see "Future improvements" in this PR description

Should we mark these as `@internal`? Or does it make more sense to export `StreamMessage` to be part of our public API because of these?  If `DecryptError` class could generate clear error messages, it maybe doesn't need to have an extra field about the raw message.

## Other packages

Updated Broker to use the new API. 

In CLI-tools there was no usages which needed updates.

## Future improvements

- We could add more fields to `Message` if needed. Currently it contains a good subset of `StreamMessage` fields, but could check if there are some fields which we should add.
  - could implement `MessageStream.test.ts` test case so that so that it doesn't call `convertStreamMessageToMessage`? (or add unit test to it) 
- StreamMessage is currently used as a property of DecryptError and other classes which extend StreamMessageError. If we want to avoid exposing StreamMessage as part of our public API we should maybe provide the data about the stream/message in Error#message property and remove the field. This could be done e.g. when we refactor all our errors to extend `StreamrClientError`.