test(client): Optimize integration tests (#969)
Optimized three integration tests. Execution time decreased:
- `Subscriber2:` 30s -> 7s
- `Resends2`: 27s -> 10s
- `GapFill`: 15s -> 8s

In `Subscriber2` and `Resends2` we are now using public streams instead of private streams. That helps us also to debug tests in the future as the public stream flow is simpler than private stream flow. As the tests case are not testing encryption related functionality, it is ok to simplify tests this way. Encryption functionality is covered by other unit, integration and end-to-end tests.

In `GapFill` test one of the test cases used 20 gap fill requests. Decreased that to use only 2 gap fill requests.

Also refactored all these three tests so that the tests are more maintainable.