refactor(broker): Extract `$schema` from `Config` (#996)
Move `$schema` property from `Config` interface to new `ConfigFile` interface. It is only needed when config files are handled.

The property is now optional. If schema is not given, no migration is done for the config file (as version info is encoded to the schema string and therefore the version is unknown).

Also removed support for testnet3 config migration.