refactor(client): Simplify `RSAKeyPair` (#999)
Simplified internal object creation logic in `RSAKeyPair`.

All components which used `RSAKeyPair` created it by calling `RSAKeyPair.create()`. In this PR we annotated the constructor as private to enforce that logic. After that enforcement we can be sure that any `RsaKeyPair` instance is always "ready"  when it is given to the caller component. Therefore:
- we don't need a separate `InitializedRSAKeyPair` type 
- we don't need `isReady` method, and calls to that method
- `publicKey` and `privateKey` fields can't be undefined as we set the during the initialization

Refactored the factory methods `keyPairServer` and `keyPairBrowser` to return `RSAKeyPair` instances instead of modifying an existing instance. Also renamed the methods to make it clear that those are environment-specific variants of the same functionality. 

Removed caching of the factory promise `(_generateKeyPairPromise`). No need to have caching as the factory is called only once for each instance.

Annotated `publicKey` and `privateKey` fields private as those are always accessed via the accessors.

Annotated return type of `getSubtle()` and removed an unnecessary null check related to that.

## Future improvements

- After this PR the only place where we use `ts-toolbelt` library is used only for MaybeAsync helper type. The dependency could be dropped if we inline/remove `MaybeAsync`.