fix(broker): Remove tracker bin (#1002)
Removed "tracker" bin script which used a `tracker.js` file via a symbolic link to `network-tracker` package. 

The `tracker.js` file was not included in the `npm` packaging (and should not, as it is in a separate package), and therefore `npm` installs for `streamr-broker` fail with error:

```
ENOENT: no such file or directory, chmod '/foo/bar/node_modules/streamr-broker/bin/tracker.js'
```

## Future improvements

Currently `run-all.js` access `tracker.js` using relative path to `network-tracker` package. That is maybe ok as it is in same monorepo. But it would be better if we could access that e.g. via `network-tracker`'s bin script? Maybe we should move this script to `monorepo` root, e.g. to new `scripts` directory there?