refactor(client ): Deprecate config defaults (#1005)
Deprecated `STREAM_CLIENT_DEFAULTS` constant. No need to expose to end-users.

Deprecated `ConfigTest` constant as it is now renamed to `CONFIG_TEST`. Currently we need to export this so that it is easy to define that a client uses the test environment.

## Future improvements

The file `ConfigTest.ts` could have better name?