feat(client): Redacted config (#1012)
The client config redacts the private key from the config object as soon as it is no longer needed. The information is available briefily in a local variable when we create an `Authentication` object, and after that we overwrite the private key with `(redacted)` string. The `StreamrClient.config` field now stores that redacted config instead of the original config.

As `StreamrClient.config` was and is a private field. Therefore this is not a functionality change of the public API.

## Refactoring

Inlined the functionality of `initContainer` so that we can creating an authentication and injecting it to the container are separated steps. Also refactored Stream unit test to use `StreamFactory` instead of `initContainer`.