refactor(client): Inline network config (#1024)
Inlined `NetworkNodeOptions` interface to `StrictStreamrClientConfig` so that it is easier to see what all the possible network config options for a `StreamrClient` instance. This way the type definition is more straightforward and therefore more readable e.g. in API reference. 

The network settings of `StreamrClientConfig` are quite similar to the `NetworkNodeOptions` interface (which is defined in `network` package), but the end-user config interface is maybe a good boundary where we explicitly state what field of that interface we use in a client. E.g. if we'd rename a field from `NetworkNodeOptions` we'd need to do some deprecation process (if we plan to release a new minor version), and this boundary enforces us to do so.

## Related changes

- Moved `disconnectionWaitTime` default value from `NetworkNodeFacade` to schema
- Annotated `acceptProxyConnections`, `iceServers` and  `disconnectionWaitTime` as required in `StrictStreamrClientConfig` as those options get default values from the schema