refactor(client): Simple config type definition (#1032)
Before this PR the main config type `StreamrClientConfig` was derived from `StrictStreamrClientConfig`. Therefore it was not trivial to see the actual field definitions e.g. in API documentation.

Now we derive the types in the opposite way: `StreamrClientConfig` has the actual fields, and `StrictStreamrClientConfig` is a stricter (i.e. many fields required) version of that type.

Also fixed the `cache` and `_timeouts` definitions so that it is possible to define those blocks partially (before this PR we needed to provide both `maxAge` and `maxSize` if we defined `cache` config).

## Dependency

Added `ts-essentials` TypeScript utility library as dev dependency. We use `MarkOptional` and `DeepRequired` types from the library. 

Another possible TypeScript library is `type-fest`. It has significantly more weekly downloads compared to `ts-essentials`. We could later use that additionally/alternatively. It doesn't currently provide `DeepRequired` utility type (but maybe it will, as it already provides e.g. `PartialDeep`).

## Future improvements

We have custom the `MaybeAsync` utility type in `types.ts`. Both `ts-essentials` and `type-fest` have a potential replacement type for that.