refactor: `run-brokers-and-trackers` script (#1040)
Refactor a script which is used to run all Brokers and Trackers in `streamr-docker-dev` environment. That script is used by CI (the `run-brokers-and-trackers` action), and internally to run manual tests.

The script was previously in `packages/broker/bin/run-all.js`. Now it is re-implemented as a Bash script in monorepo root `bin/run-brokers-and-trackers.sh`. The monorepo root is a better location for the script because it uses Trackers (which is not even a dependency of `broker` package).

## Other changes

- Removed obsolete `tracker` script from `network-tracker`: no need to have a separate `npm` script because we can start the bin with `npm exec`
- Removed obsolete `packages/broker/bin/kill-all.sh`