fix: Update package lock (#1043)
Update `package-lock.json` file. Most likely the lock file which was included in https://github.com/streamr-dev/network/pull/1028 was partially cached and therefore was not fully updated.

Fixed also build and eslint issues related to that TypeScript update PR. It is not clear why `eslint-disable-next-line promise/no-promise-in-callback` is needed in `SubscriptionSession.ts:84` (e.g. there is similar implementation in line 92, which doesn't need eslint ignore).