fix(client): `allOf` filters in `searchStreams` (#1049)
Fixed the query for search stream functionality. If `allOf` filter is used to search streams, the result set didn't include streams which have `edit`, `delete` or `grant` permission granted unless those permissions are also included in the `allOf` filter.

The GraphQL query should have excluded the permissions which are not in the `allOf` filter by defining the permission query fields as `undefined`. For `publish` and `subscribe` permissions it was handled correctly. For `edit`, `delete` or `grant` the permission fields were incorrectly set to `false` instead of `undefined`.

## Test

No unit or end-to-end test caught this bug. Could consider adding a test.