fix(utils): Duplicate invocations in `scheduleAtFixedRate` (#1074)
Bug fix to `scheduleAtFixedRate`.

### Previous functionality

It seems that before this fix the scheduler invoked the callback twice in some rare situations. We were not able to reproduce that in our environments.  Instead we have observed a bug that some Broker nodes published metrics samples twice for some periods.  

The first of those duplicated samples was corrupted: it did not contain any `RateMetric` fields. In `RateMetric` the values are `undefined` if the start and stop times are the same. The reason for the identical timestamp is most likely that the callback in `MetricsContext#createReportProducer` the callback of `scheduleAtFixedRate` was sometimes called twice for the same `now` value.

The previous implementation assumed that when we:
- run a task at `T0`, which takes `s1` milliseconds
- and then we wait for (`interval` - `s1` milliseconds)
- then there is at least an `interval` milliseconds delay between `T0` and `T1`

That the assumption is not always valid, see e.g. https://github.com/streamr-dev/network/blob/main/packages/utils/test/wait.test.ts#L12

### New functionality

The new implementation ensures that we don't invocate a task multiple times for any `now` value. There is now an explicit variable which stores the timestamp of current invocation. We increment that value explicitly instead of calculating the new value from the current `now` value.