fix(broker): Do not wait for server close event (#1083)
Before this PR if a connected client didn't respond to server-initiated close, the server wasn't able to stop. It waited for callback from `this.wss.close()` . That callback is not called if the client doesn't respond to that action.

E.g. if a client is paused (`client.pause()`), it doesn't send any response to `this.wss.close()` and therefore we can't wait for such clients to respond before stopping the server.

The issue is solved in this PR by ignoring the callback. The callback is redundant as we already explicitly terminate each connection by calling `ws.terminate()`.