ci(deps): bump jest from 28.1.2 to 29.4.3 (#1086)
Updated `jest`, `jest-mock`, `jest-extended` and `ts-jest`.

The `jest-leak-detector` was not updated. With `29.4.3` `MemoryLeaks.test.ts` fails. The test is maybe flakey, or maybe our custom `LeaksDetector` is not compatible with the new version, or maybe the new version was able to find a new memory leak. When we update to `29.4.3` we can most likely also remove the `weak-napi` dependency (https://github.com/facebook/jest/blob/main/CHANGELOG.md#2900).

Updated also `@types/merge2`