NET-943:  Fallback proxies (#1073)
## Summary

This is a breaking change as it changes the client's public API! 

The proxy pattern is changed to function based on selecting `k` nodes to connect to out of `n` candidates. The client API is changed as a result of the changes. However, new clients should be able to open proxy connections to old clients.

## Changes

- The ProxyConnectionManager is split to two separate files for the client and server sides.
- If a disconnection on a Proxy Connection occurs, the client will select a new random known candidate node and attempt to open a Proxy Connection to it.
- The `k` value an be increased and reduced after candidates have been set. The node will attempt to connect up to or disconnect down to the new number of target connections.

## Limitations and future improvements

Directionality (publish / subscribe ) is now only set when candidates are set. When candidates are removed the candidates are not differentiated based on direction. (It was never the intention that nodes should be able to open proxy connections to both directions on a single stream. In this case the full node pattern should be used)