Improve docker tags (#1114)
## Summary

Change the way broker docker image tags are published to Docker Hub. We want the following rules to be observed:
- When pushing a git tag, e.g.`broker/v31.8.9` then the equivalent broker docker image is published under `v31.8.9` tag to Docker Hub.
- When pushing anything into `main` branch, the equivalent broker docker image get published to Docker Hub with tag `dev`.
- When pushing a git tag, e.g.`broker/v31.8.9` that *does not* contain words `beta` or `alpha`, then the equivalent broker docker image is published under tag `latest`.

Notice that we may publish a version tag and `latest` tag in one go.

## Limitations and future improvements

There is one particularly annoying edge case that may lead to unintended `latest` tags.

Say you have published a version `31.8.9` and then later decide to publisher `32.0.0`. Now if you decide you want to release a patch version `31.8.10`, it will become `latest`. This is most likely not be what is wanted as the intention of `latest` would be to have the very recent latest major version. This needs to be addressed in a later PR, probably in the near future.