feat: prohibit setProxies when acceptProxyConnections=true (#1156)
## Summary

Prevent the use of method `networkNode#setProxies` when accepting proxy connections (i.e. `acceptProxyConnections=true`). Throw a human readable error if such usage is attempted.

## Limitations and future improvements

The main reason for implementing this restriction is due to incomplete state management in the code when a node
- wishes to act as a proxy themselves, and
- wishes to subscribe / publish to a stream using another proxy node.

Our evaluation is that this is a very unlikely scenario in practice. If such behavior is wanted in the future, the restriction imposed in this PR should be removed and the state management logic should be changed to accommodate this scenario.