chore: remove exported validateConfig (#1166)
## Summary

Remove exported function `validateConfig` from client. It was used by broker to validate config but this is done within `StreamrClient` constructor as well so no need to double validate. (There used to be a good reason to do the validation in broker.)

## Limitations and future improvements

This is technically a breaking change with regards to semver; in practice this function is used internally only. So will be a bit naughty here and not increase major version.