test(client): NET-956 Check docs (#1167)
Run `npm run check` in CI to check if there are any warnings when we generate API docs with `typedoc`. A warning typically means that we have not exported some type/class/interface should be part of the public API.

Also updated `typedoc` as the previous version didn't fully support TypeScript 4.9.